back_end/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
motor==3.3.1
pymongo==4.6.0
python-dotenv==1.0.0
email-validator==2.1.0
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4
# 用于生产环境
gunicorn==21.2.0 